Move the code editor filter to its own file
