+++ inc/Editor/CrdtPersistence.php
@@ -2,8 +2,6 @@
 
 namespace VIPRealTimeCollaboration\Editor;
 
-use WP_Block_Editor_Context;
-
 use VIPRealTimeCollaboration\Compatibility\Compatibility;
 use function add_action;
 use function register_meta;
@@ -19,7 +17,6 @@ final class CrdtPersistence {
 
 	public function __construct() {
 		add_action( 'init', [ $this, 'register_meta' ], 999, 0 );
-		add_filter( 'block_editor_settings_all', [ $this, 'disable_code_editor' ], 10, 2 );
 	}
 
 	public function register_meta(): void {
@@ -47,32 +44,5 @@ final class CrdtPersistence {
 				]
 			);
 		}
-	}
-
-	/**
-	 * Disable the code editor.
-	 *
-	 * @psalm-suppress PossiblyUnusedReturnValue Psalm does not detect usage via add_filter.
-	 */
-	public function disable_code_editor( array $settings, WP_Block_Editor_Context $context ): array {
-		// Only modify settings for the post editor context.
-		if ( ! $context->name || ! $context->post || 'core/edit-post' !== $context->name ) {
-			return $settings;
-		}
-
-		// Get the post, and supported post types for collaboration from the arguments provided.
-		$supported_post_types = Compatibility::get_supported_post_types();
-		$post = $context->post;
-
-		// If the post doesn't support collaboration, return the original settings.
-		if ( ! in_array( $post->post_type, $supported_post_types, true ) ) {
-			return $settings;
-		}
-
-		// Disable switching to code editing mode.
-		$settings['codeEditingEnabled'] = false;
-
-		// Return the modified settings.
-		return $settings;
 	}
 }
