+++ vip-real-time-collaboration.php
@@ -17,6 +17,7 @@ use VIPRealTimeCollaboration\Api\RestApi;
 use VIPRealTimeCollaboration\Assets\Assets;
 use VIPRealTimeCollaboration\Auth\SyncPermissions;
 use VIPRealTimeCollaboration\Compatibility\Compatibility;
+use VIPRealTimeCollaboration\Editor\EditorUI;
 use VIPRealTimeCollaboration\Editor\CrdtPersistence;
 use VIPRealTimeCollaboration\Overrides\Overrides;
 
@@ -53,6 +54,7 @@ add_action( 'plugins_loaded', static function (): void {
 
 	new Assets();
 	new Compatibility();
+	new EditorUI();
 	new CrdtPersistence();
 	new Overrides();
 	new RestApi();
